refactor : 하드코딩된 String 문자열 상수화
